Changes for 64-bit compatibility by Brian Bergstrand: Keyed archiving; 64-bit upgrade of NSArchiver data; upgrade sparkle version; upgrade Xcode project version to 3.1; targeting 10.5 SDK instead of 10.4. Also increase resiliency against potential serialization problems by verifying writes before saving; allow recovery of notes in the case that the database cannot be rewritten

+++ DeletionManager.m
@@ -188,19 +188,19 @@
 	}
 }
 
-- (BOOL)tableView:(NSTableView *)aTableView shouldEditTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (BOOL)tableView:(NSTableView *)aTableView shouldEditTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	return NO;
 }
 
-- (BOOL)tableView:(NSTableView *)aTableView shouldSelectRow:(int)rowIndex {
+- (BOOL)tableView:(NSTableView *)aTableView shouldSelectRow:(NSInteger)rowIndex {
 	return NO;
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	return filenameOfNote((NoteObject *)[deletedNotes objectAtIndex:rowIndex]);
 }
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView {
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView {
     return [deletedNotes count];
 }
