+++ FastListDataSource.h
@@ -11,25 +11,25 @@
 
 @interface FastListDataSource : NSObject {
 	id *objects;
-    unsigned int count;
+    NSUInteger count;
 	IMP objRetain, objRelease;
 }
 
 - (id)initWithClass:(Class)aClass;
 - (const id *)immutableObjects;
-- (unsigned int)count;
+- (NSUInteger)count;
 
-- (unsigned)indexOfObjectIdenticalTo:(id)address;
+- (NSUInteger)indexOfObjectIdenticalTo:(id)address;
 - (NSArray*)objectsAtFilteredIndexes:(NSIndexSet*)indexSet;
 
 - (void)fillArrayFromArray:(NSArray*)array;
 - (BOOL)filterArrayUsingFunction:(BOOL (*)(id, void*))present context:(void*)context;
 
-- (void)sortStableUsingFunction:(int (*)(id *, id *))compare;
+- (void)sortStableUsingFunction:(NSInteger (*)(id *, id *))compare;
 
 - (void)tableView:(NSTableView *)aTableView setObjectValue:(id)anObject 
-   forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex;
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex;
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView;
+   forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex;
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex;
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView;
 
 @end
